added factory folder for addresses, partners and users for simplyfied use
